backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
influxdb-client==1.40.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
